tutorials/13-full-stack-applications/ai-stylist/requirements.txt
streamlit
git+https://github.com/huggingface/transformers
image
langchain
botocore
langchain-ibm
ibm-watsonx-ai
python-dotenv
torch
markupsafe
